phase 1

+++ Member.cs
@@ -9,7 +9,7 @@ namespace heist
         public int Skill { get; set; }
         public decimal Courage { get; set; }
 
-        public Member(string name, int skill, int courage)
+        public Member(string name, int skill, decimal courage)
         {
             Name = name;
             Skill = skill;
